refactor: extract super ops subprocessor actions
